Fixed a variety of display and grid issues
This project was the first use of grid-templating so some things were a little strange. Additionally, the SVGs used in the animations didn't have hidden overflows, so the page (especially on mobile) would push off to the right as the car SVG is supposed to move off the right side of the screen. Things are fixed up now, and the design fits the grid correctly and is lightly scalable.